Started work on the new video page
Removed old video code.  Replaced it with the start for the video page.
Right now we only render the site header, but that's fairly significant
because it requires all the CSS and JS to be bundled correctly.

We're now in a fairly weird situation with frontend code.  We have the
legacy code, the newer code, and now this code.  I put the new template
code in the "future" directory, I figure this sounds like the from
__future__ imports in python.  Called the new CSS/JS bundles "amara.css"
and "amara.js".  Hopefully all these different naming systems can
coexist.

Added some new JS/CSS dependencies.  These are managed using bower.js,
and also checked in to the repository in the media/bower directory.

I put the new source JS/CSS in media/amara.  Then it goes in the js/css
subdirectory.  The main reason I went for this is that's how it's done
with most other packages.  Also the src subdirectory didn't make much
sense to me, since we have different sources for the site, the editor,
and the embedder.