Implemented the new video page
Added a bunch of javascript for handling the forms on the page

Fix for calling send_video_comment_notification().

Changed small user thumb size to 30px.  As far as I can tell, we're not
currently using it anywhere, so this change shouldn't affect the old
styled pages.

Got the icon font working